two translation objects
the first is the standard way of updating the original `i18n` object
that is set by `Translator$new` when using a server-side UI

The second translation object is `i18n_r`, and is a reactive version of
the original `i18n`. It is used in the `update*` functions that update
the language of strings used in the widgets for the sidbar.